+++ src/tests/test.index.js
@@ -1,3 +1,4 @@
+require('hard-rejection')();
 const Eth = require('../index.js');
 const Eth2 = require('../index.js');
 const assert = require('chai').assert;
@@ -288,10 +289,10 @@ describe('ethjs-query', () => {
 
       /// @title Voting with delegation.
       contract Ballot {
-        function () payable {
+        function () public payable {
         }
 
-        uint cool;
+        uint256 public cool;
       }
       `;